requirements.txt
langchain_nvidia_ai_endpoints
python-dotenv
openai
langchain
pyyaml
tiktoken
huggingface-hub